requirements.txt
sentence_transformers
numpy
langchain
langchain-community 
chromadb 
fastapi
uvicorn
python-multipart
redis
chromadb
pypdf
python-dotenv
pika